Create a subdir for the site to avoid serving the README.
`README.md` was being renamed to `readme.md` (as seen in a
download of the released site), and that was being served if
asked for explicitly, despite the redirects file. So create
a subdir for the site with only the `_redirects` file so that
we don't inadvertantly serve anything else.