Completely redo as per azurelogic Metalsmith blog tutorial
https://azurelogic.com/posts/building-a-blog-with-metalsmith/

This currently puts the posts on the wrong paths (he uses
`/posts/post-title`.

A couple of minor changes from the code on the web page were required
to get this to work; he was using `posts/**.html` instead of
`posts/**/*.html`. This may have worked for him because he was
possibly storing his posts directly under the `posts/` directory
rather than under subdirectories of it.

------------