Move blog index from `/posts.html` to `/blog/`.
We update the pattern in the collections to specifically exclude it so
it doesn't show up in the listing of posts.

We leave `paginate: posts` in place in the metadata for it; I'm not
totally clear on how that works but I guess it refers to the `posts`
JS variable passed to `collections()` in `index.js`.