Test: Fix -C option, and add -c as well
The -C option was doing a `git clean -X` to remove all (and only)
ignored files, which is definitely *not* what anybody wants
(especially when they made a temporary copy of a file for later use).
Change this instead simply to remove .build/.

We also add a -c alias for this, since even this "big clean" is still
a very small clean (a few seconds rebuild time) for all but the
largest sites.

We now also use `site-compiler build` instead of `site-compiler
rebuild` (the latter of which does a clean build of .build/site)
because the cleaning is now taken care of by -c if the user wanted it.