Post-ideas,…/draft/configuration-is-code: Files from a branch
These are from the `next-post` branch, which branch we will no longer
be using. Since all of the branches are available to the public on
GitHub anyway, there's no point to not including /draft/… URLs on the
production website with the draft posts as well, so long as they're
not in the listings.